UI modif, add bus. Some techincalities to edit, passing params

+++ Sprint 2/BusReservationSystem/src/main/java/com/cg/busreservationsystem/service/AdminService.java
@@ -1,5 +1,6 @@
 package com.cg.busreservationsystem.service;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -9,8 +10,8 @@ import com.cg.busreservationsystem.dto.Transaction;
 public interface AdminService {
 	
 	public Bus addBusDetails(Bus bus);
-	public Integer removeBusDetails(Integer busId);
-	public Bus searchBus(Integer busId);
+	public Integer removeBusDetails(BigInteger busId);
+	public Bus searchBus(BigInteger busId);
 	public List<Bus> viewBuses();
 	public List<Transaction> getTransactionsByDate(LocalDate date);
